src/requirements.txt
backcall==0.1.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
decorator==4.4.2
dnspython==1.16.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.14.1
lxml==4.5.0
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pymongo==3.10.1
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
requests-file==1.5.1
six==1.14.0
soupsieve==2.0
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
